% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{pentagon_colours}
\alias{pentagon_colours}
\title{Pentagon Colours}
\format{
An object of class \code{list} of length 5.
}
\usage{
pentagon_colours
}
\description{
Pentagon Colours
}
\examples{
library(unikn)
seecol("pentagon_colours")

}
\keyword{datasets}
