% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{pentagon_palette}
\alias{pentagon_palette}
\title{\code{pentagon_palette}}
\usage{
pentagon_palette(
  direction = 1,
  primary = "Punky Orange",
  other = "Mulberry Wine"
)
}
\arguments{
\item{return}{will return corresponding hexcode of the colour name to the ggplot.}
}
\description{
uses the custom color palette to be used with ggplot2 to generate the colors for discrete variables with maximum of 5 categories.
\describe{
\item{Punky Orange}
\item{Gold Yellow}
\item{Dusty Green}
\item{Indigo Blue}
\item{Mulberry Wine}

}
}
\section{Arguments}{

}

