% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale1.R
\name{scale_color_pentagon_d}
\alias{scale_color_pentagon_d}
\title{Color Scales for Discrete Variables:

\code{scale_colour_pentagon_d } is a color scales used with ggplot2 for continuous data values mapped on the color aesthetic. It chooses the branded palette and the level of the color scale can be changed using the \param{direction}.}
\usage{
scale_color_pentagon_d(
  primary = "Punky Orange",
  other = "Mulberry Wine",
  direction = 1,
  ...
)
}
\arguments{
\item{primary}{a character vector to choose a color from the palette}

\item{other}{numeric values to set the variation in color between high and low scale}
}
\description{
Color Scales for Discrete Variables:

\code{scale_colour_pentagon_d } is a color scales used with ggplot2 for continuous data values mapped on the color aesthetic. It chooses the branded palette and the level of the color scale can be changed using the \param{direction}.
}
\details{
Default colours are generated with branded palette. Generally, for continuous colour scales you want to keep hue constant, but vary chroma and luminance.
}
\section{Arguments}{
NA
}

\examples{

df <- data.frame(
x = runif(100),
y = runif(100),
z1 = rnorm(100),
z2 = abs(rnorm(100))
)



# For coloring the based on continuous variable
ggplot(df, aes(x, y)) +
  geom_point(aes(colour = z1)) +
    scale_colour_pentagon_d ()

# For coloring the based on continuous variable on a reverse scale
ggplot(df, aes(x, y)) +
  geom_point(aes(colour = z1)) +
    scale_colour_pentagon_d (direction=-1)

}
\seealso{
\code{\link[ggplot2]{scale_color_gradient}}
}
