% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-scale.R
\name{scale_fill_pentagon_d}
\alias{scale_fill_pentagon_d}
\title{Colour Scale Fill for Discrete Variables :

\code{scale_fill_pentagon_d }}
\usage{
scale_fill_pentagon_d(
  primary = "Punky Orange",
  other = "Mulberry Wine",
  direction = 1,
  ...
)
}
\arguments{
\item{...}{Other parameters that support the scale_colour_discrete function.}
}
\description{
It is a color scales used with ggplot2 to fill discrete data values mapped on the color aesthetic of Visualization in R. It chooses the Pentagon Palette which consists of 5 colors:
\describe{
\item{Punky Orange}
\item{Gold Yellow}
\item{Dusty Green}
\item{Indigo Blue}
\item{Mulberry Wine}

}
The level of the color scale can be changed using the \code{direction}.
}
\details{

}
\examples{

# Data Used
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]

# Basic Code
 ggplot(dsamp, aes(cut, price, fill = cut)) +
     geom_col()+
     scale_fill_pentagon_d()


# Direction Reversed
 ggplot(dsamp, aes(cut, price, fill = cut)) +
     geom_col()+
     scale_fill_pentagon_d(direction = -1)


}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
