% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R, R/pentagon_colours.R
\docType{data}
\name{pentagon_colours}
\alias{pentagon_colours}
\title{Pentagon Color Palette:

Pentagon Colours}
\format{
An object of class \code{list} of length 5.

An object of class \code{list} of length 5.
}
\usage{
pentagon_colours

pentagon_colours
}
\description{
Pentagon Color Palette:

Pentagon Colours

Pentagon Colours
}
\examples{
pentagon_colours
pentagon_colours
}
\keyword{datasets}
