% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{pentagon_colours}
\alias{pentagon_colours}
\title{Pentagon Color Palette:

\code{pentagon_palette} uses the custom color palette to be used wuth ggplot2 to generate the colors for discrete variables with maximum of 5 categories.}
\format{
An object of class \code{list} of length 5.
}
\usage{
pentagon_colours
}
\arguments{
\item{direction:}{a numeric value to set the direction of the color scale \newline if direction is < 0 the color list is reversed}
}
\description{
Pentagon Color Palette:

\code{pentagon_palette} uses the custom color palette to be used wuth ggplot2 to generate the colors for discrete variables with maximum of 5 categories.
}
\details{
Default colors are generated using this branded palette.
}
\section{Arguments}{

}

\section{primary the baseline color for continuos scale}{
NA
}

\seealso{
\code{\link{pentagon_palette}} \code{\link[ggplot2]{discrete_scale}} \code{\link[ggplot2]{scale_colour_pentagon_seq_c}}
}
\keyword{datasets}
