% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale1.R
\name{scale_color_pentagon_d}
\alias{scale_color_pentagon_d}
\title{Color Scales for Discrete Variables:

\code{scale_colour_pentagon_d }}
\usage{
scale_color_pentagon_d(
  primary = "Punky Orange",
  other = "Mulberry Wine",
  direction = 1,
  ...
)
}
\arguments{
\item{...}{Other parameters that support the scale_colour_discrete function}
}
\description{
It is a color scales used with ggplot2 for discrete data values mapped on the color aesthetic of Visualization in R. It chooses the Pentagon Palette which consists of 5 colors:
\describe{
\item{Punky Orange}
\item{Gold Yellow}
\item{Dusty Green}
\item{Indigo Blue}
\item{Mulberry Wine}

}
The level of the color scale can be changed using the \code{direction}.
}
\examples{
# Data Used
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]

# Basic Code
ggplot(dsamp, aes(carat, price)) +
  geom_point(aes(colour = cut), size = 3)+
  scale_color_pentagon_d()

# Direction reversed
ggplot(dsamp, aes(carat, price)) +
  geom_point(aes(colour = cut), size = 3)+
  scale_color_pentagon_d(direction = -1)


}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
