% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale1.R
\name{scale_color_pentagon_seq_c}
\alias{scale_color_pentagon_seq_c}
\title{Color Scales for Continuous Variables:}
\usage{
scale_color_pentagon_seq_c(
  name = "Mulberry Wine",
  direction = 1,
  amount = 0.6,
  ...
)
}
\arguments{
\item{name}{a character vector to choose a color from the palette. Default value is Mulberry Wine}

\item{direction}{a numeric value to set the level of the scale. If direction is < 0 , the palette switches from high and low}

\item{amount}{a double vector with which the amount of opacity of the color can be adjusted. Default value is 0.6}

\item{...}{Other parameters that support the scale_colour_discrete function}
}
\description{
\code{scale_colour_pentagon_seq_c } It is a color scales used with ggplot2 for continuous data values mapped on the color aesthetic of Visualization in R. It chooses the Pentagon Palette which consists of 5 colors. The level of the color scale can be changed using the \code{direction}.
}
\details{
Default colour gradients are generated with pentagon palette. Generally, for continuous colour scales you want to keep hue constant, but vary chroma and luminance.
}
\examples{

# Data Used
library(ggplot2)
library(dye)
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]

# Basic continuous variable
ggplot(dsamp, aes(price, depth, color = price)) +
 geom_point( size = 3)+
  scale_color_pentagon_seq_c() +
  ggtitle("Example of Continuous Scale")


# Basic continuous variable with Different Colour
ggplot(dsamp, aes(price, depth, color = price)) +
 geom_point( size = 3)+
  scale_color_pentagon_seq_c("Indigo Blue") +
  ggtitle("Example of Continuous Scale with Different Colour")


# Basic continuous variable on a Reverse Scale
ggplot(dsamp, aes(price, depth, color = price)) +
 geom_point( size = 3) +
  scale_colour_pentagon_seq_c(direction = -1) +
  ggtitle("Example of Continuous Scale Reverse")


# Basic continuous variable on a Reverse Scale with Different Colour
ggplot(dsamp, aes(price, depth, color = price)) +
 geom_point( size = 3)+
  scale_colour_pentagon_seq_c("Dusty Green", direction = -1) +
  ggtitle("Example of Continuous Scale Reverse with Different Colour")

}
\seealso{
\code{\link[ggplot2]{scale_color_gradient}}
}
