% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale1.R
\name{scale_color_pentagon_seq_c}
\alias{scale_color_pentagon_seq_c}
\title{Color Scales for Continuous Variables:

\code{scale_colour_pentagon_seq_c }}
\usage{
scale_color_pentagon_seq_c(
  name = "Mulberry Wine",
  direction = 1,
  amount = 0.6,
  ...
)
}
\description{
It is a color scales used with ggplot2 for continuous data values mapped on the color aesthetic of Visualization in R. It chooses the Pentagon Palette which consists of 5 colors:
\describe{
\item{Punky Orange}
\item{Gold Yellow}
\item{Dusty Green}
\item{Indigo Blue}
\item{Mulberry Wine}

}
The level of the color scale can be changed using the \code{direction}.
}
\examples{

# Data Used
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]

# Basic continuous variable
ggplot(dsamp, aes(price, depth, color = price)) +
 geom_point( size = 3)+
  scale_color_pentagon_seq_c() +
  ggtitle("Example of Continuous Scale")


# Basic continuous variable with Different Colour
ggplot(dsamp, aes(price, depth, color = price)) +
 geom_point( size = 3)+
  scale_color_pentagon_seq_c("Indigo Blue") +
  ggtitle("Example of Continuous Scale with Different Colour")


# Basic continuous variable on a Reverse Scale
ggplot(dsamp, aes(price, depth, color = price)) +
 geom_point( size = 3) +
  scale_colour_pentagon_seq_c(direction = -1) +
  ggtitle("Example of Continuous Scale Reverse")


# Basic continuous variable on a Reverse Scale with Different Colour
ggplot(dsamp, aes(price, depth, color = price)) +
 geom_point( size = 3)+
  scale_colour_pentagon_seq_c("Dusty Green", direction = -1) +
  ggtitle("Example of Continuous Scale Reverse with Different Colour")

}
\seealso{
\code{\link[ggplot2]{scale_color_gradient}}
}
