% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-scale.R
\name{scale_fill_pentagon_d}
\alias{scale_fill_pentagon_d}
\title{Colour Scale Fill for Discrete Variables :

\code{scale_fill_pentagon_d } is a color scales used with ggplot2 for continuous data values mapped on the color aesthetic. It chooses the branded palette and the level of the color scale can be changed using the \param{direction}.}
\usage{
scale_fill_pentagon_d(
  primary = "Punky Orange",
  other = "Mulberry Wine",
  direction = 1,
  ...
)
}
\arguments{
\item{primary}{a character vector to choose a color from the palette}

\item{other}{numeric values to set the variation in color between high and low scale}

\item{direction:}{a numeric value to set the level of the scale \newline if direction is < 0 switch high and low}
}
\description{
Colour Scale Fill for Discrete Variables :

\code{scale_fill_pentagon_d } is a color scales used with ggplot2 for continuous data values mapped on the color aesthetic. It chooses the branded palette and the level of the color scale can be changed using the \param{direction}.
}
\details{
Default colours are generated with branded palette. Generally, for continuous colour scales you want to keep hue constant, but vary chroma and luminance.
}
\section{Arguments}{
NA
}

\examples{

df <- data.frame(
x = runif(100),
y = runif(100),
z1 = rnorm(100),
z2 = abs(rnorm(100))
)



# For coloring the based on continuous variable
ggplot(df, aes(x, y)) +
  geom_point(aes(colour = z1)) +
    scale_fill_pentagon_d ()

# For coloring the based on continuous variable on a reverse scale
ggplot(df, aes(x, y)) +
  geom_point(aes(colour = z1)) +
    scale_fill_pentagon_d (direction=-1)

}
\seealso{
\code{\link[ggplot2]{scale_color_gradient}}
}
